TQAunit mistake container was extended to 32 bit to support more kinds of mistakes. All dependices adjusted to it. Reading 8bit from old binary exam files is still supported.

+++ src/libs/core/exam/tattempt.cpp
@@ -26,7 +26,7 @@ Tattempt::Tattempt() :
 }
 
 
-void Tattempt::add(quint8 mistake, quint32 time) {
+void Tattempt::add(quint32 mistake, quint32 time) {
 	mistakes << mistake;
 	if (time)
 		times << time;
